1st change to merge searchbox to books section
