+++ package.json
@@ -23,7 +23,7 @@
     "@react-native-firebase/app": "^17.4.2",
     "@react-native-firebase/auth": "^17.4.2",
     "@react-native-firebase/crashlytics": "^17.4.2",
-    "@react-native-firebase/messaging": "^21.4.1",
+    "@react-native-firebase/messaging": "^21.5.0",
     "@react-navigation/bottom-tabs": "^7.0.0",
     "@react-navigation/native": "^7.0.0",
     "@react-navigation/native-stack": "^7.0.0",
